requirements.txt
django
djangorestframework
markdown
django-filter
psycopg2
colander
coreapi
-e .